+++ colossalai/fx/tracer/bias_addition_patch/patched_bias_addition_module/__init__.py
@@ -1,2 +1,3 @@
+from .bias_addition_module import *
 from .conv import *
 from .linear import *
