+++ colossalai/fx/tracer/meta_patch/__init__.py
@@ -1,5 +1,2 @@
-from .patched_bias_addition_function import *
-from .patched_bias_addition_module import *
 from .patched_function import *
 from .patched_module import *
-from .registry import *
