+++ colossalai/fx/tracer/meta_patch/patched_module/normalization.py
@@ -1,5 +1,6 @@
 import torch
-from ..registry import meta_patched_module
+
+from ...registry import meta_patched_module
 
 
 @meta_patched_module.register(torch.nn.LayerNorm)
